improve `NestedTest`
